feat: Refactor SRS protocol and server for improved type safety
Updated protobuf definitions to use more precise types (e.g., float32 for frequencies, uint32 for radio IDs), removed unnecessary request parameters, and restructured server-side logic to use UUIDs and context values for client identification. Improved coalition and radio handling, added missing utility functions, and enhanced error handling and logging throughout the authentication and radio services. These changes improve type safety, maintainability, and protocol clarity.

+++ utils/auth.go
@@ -186,7 +186,7 @@ func getJWTClaims(tokenString, privateKeyFile, publicKeyFile string) (*TokenClai
 	if claims, ok := token.Claims.(jwt.MapClaims); ok && token.Valid {
 		tokenClaims := &TokenClaims{
 			ClientGuid: claims["client_guid"].(string),
-			RoleId:     claims["role_id"].(uint8),
+			RoleId:     uint8(claims["role_id"].(float64)),
 		}
 		return tokenClaims, nil
 	}
